+++ app/src/main/java/sugtao4423/lod/App.kt
@@ -8,6 +8,7 @@ import sugtao4423.lod.db.UseTimeRoomDatabase
 import sugtao4423.lod.entity.Account
 import sugtao4423.lod.model.*
 import sugtao4423.lod.service.AutoLoadTLService
+import sugtao4423.lod.utils.ShowToast
 import twitter4j.StatusUpdate
 import twitter4j.Twitter
 import twitter4j.TwitterException
