+++ app/src/main/java/sugtao4423/lod/dialog_onclick/Dialog_favorite.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import sugtao4423.lod.App
 import sugtao4423.lod.R
-import sugtao4423.lod.ShowToast
+import sugtao4423.lod.utils.ShowToast
 import twitter4j.Status
 import twitter4j.TwitterException
