+++ app/src/main/java/sugtao4423/lod/ui/addaccount/AddAccountActivityViewModel.kt
@@ -17,7 +17,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import sugtao4423.lod.App
 import sugtao4423.lod.R
-import sugtao4423.lod.ShowToast
+import sugtao4423.lod.utils.ShowToast
 import sugtao4423.lod.entity.Account
 import twitter4j.Twitter
 import twitter4j.TwitterFactory
