+++ app/src/main/java/sugtao4423/lod/ui/settings/Follow2ListViewModel.kt
@@ -10,7 +10,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import sugtao4423.lod.App
 import sugtao4423.lod.R
-import sugtao4423.lod.ShowToast
+import sugtao4423.lod.utils.ShowToast
 import twitter4j.ResponseList
 import twitter4j.UserList
