+++ app/src/main/java/sugtao4423/lod/ui/showimage/ShowImageActivityViewModel.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.LiveData
 import com.hadilq.liveevent.LiveEvent
 import sugtao4423.lod.App
 import sugtao4423.lod.R
-import sugtao4423.lod.ShowToast
+import sugtao4423.lod.utils.ShowToast
 import sugtao4423.lod.utils.Regex
 
 class ShowImageActivityViewModel(application: Application) : AndroidViewModel(application) {
