+++ app/src/main/java/sugtao4423/lod/ui/userpage/fragment/DetailFragmentViewModel.kt
@@ -12,7 +12,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import sugtao4423.lod.App
 import sugtao4423.lod.R
-import sugtao4423.lod.ShowToast
+import sugtao4423.lod.utils.ShowToast
 import twitter4j.User
 
 class DetailFragmentViewModel(application: Application) : AndroidViewModel(application) {
