+++ app/src/main/java/sugtao4423/lod/ui/userpage/fragment/UserFragmentViewModel.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import sugtao4423.lod.R
-import sugtao4423.lod.ShowToast
+import sugtao4423.lod.utils.ShowToast
 import sugtao4423.lod.ui.BaseTweetListViewModel
 import twitter4j.PagableResponseList
 import twitter4j.User
